server/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
torch
pandas
numpy

# create vitural env
# python -m venv virtenv 
# activate deactivate virtenv
# source virtenv/bin/activate
# deactivate

# create a django project
# django-admin startproject core .
# this contains all the django boilerplate
# create a new app 
# python manage.py startapp api